Split example app by API version
